Move SDL directory out of the way.
